externalize importer logic from the view
